+++ RELEASE-NOTES.md
@@ -5,6 +5,7 @@
 * #48 Combine bootique-jersey with bootique-jersey-client in one project 
 * #49 Upgrade to Jersey 2.30.1
 * #50 Integrate 'jersey-bean-validation'
+* #51 Resources with dynamically defined paths
 
 ## 1.0.RC1